Move generic impls of NumOpResult<T>
The generic impls should live in the same file as the type instead of
nested in the `amount` module because they are not specific to `amount`.

Internal, no logic change.
