+++ units/src/amount/result.rs
@@ -187,57 +187,6 @@ crate::internal_macros::impl_op_for_references! {
 
         fn rem(self, modulus: i64) -> Self::Output { self.and_then(|lhs| lhs % modulus) }
     }
-
-    impl<T> ops::Add<NumOpResult<T>> for NumOpResult<T>
-    where
-        (T: Copy + ops::Add<Output = NumOpResult<T>>)
-    {
-        type Output = NumOpResult<T>;
-
-        fn add(self, rhs: Self) -> Self::Output {
-            match (self, rhs) {
-                (R::Valid(lhs), R::Valid(rhs)) => lhs + rhs,
-                (_, _) => R::Error(NumOpError::while_doing(MathOp::Add)),
-            }
-        }
-    }
-
-    impl<T> ops::Add<T> for NumOpResult<T>
-    where
-        (T: Copy + ops::Add<NumOpResult<T>, Output = NumOpResult<T>>)
-    {
-        type Output = NumOpResult<T>;
-
-        fn add(self, rhs: T) -> Self::Output { rhs + self }
-    }
-
-    impl<T> ops::Sub<NumOpResult<T>> for NumOpResult<T>
-    where
-        (T: Copy + ops::Sub<Output = NumOpResult<T>>)
-    {
-        type Output = NumOpResult<T>;
-
-        fn sub(self, rhs: Self) -> Self::Output {
-            match (self, rhs) {
-                (R::Valid(lhs), R::Valid(rhs)) => lhs - rhs,
-                (_, _) => R::Error(NumOpError::while_doing(MathOp::Sub)),
-            }
-        }
-    }
-
-    impl<T> ops::Sub<T> for NumOpResult<T>
-    where
-        (T: Copy + ops::Sub<Output = NumOpResult<T>>)
-    {
-        type Output = NumOpResult<T>;
-
-        fn sub(self, rhs: T) -> Self::Output {
-            match self {
-                R::Valid(amount) => amount - rhs,
-                R::Error(_) => self,
-            }
-        }
-    }
 }
 
 impl_mul_assign!(NumOpResult<Amount>, u64);
